Abstract out tabs
